==> analyses/1. generate placebo data.rmd <==
```{r}
options(digitis = 3)
options(scipen = 10^5)
options(help_type = "html")

library(parallel)
library(tidyverse)

#load R functions
sapply(list.files("R/", ".r", full.names = TRUE), source)
library(data.table)

library(rpart)
library(ranger)
```

Select data
(compact is created in 3.0, basically a subset)
```{r}
dt <-  readRDS("data/mock_dt.rds")
dt <- as.data.table(dt, key = c("i", "year"))
dt <- reset_y0(dt)
summary(dt)
summary(dt$start_year)

table(dt$start_year)

```

The propesity os estimated only once based on actual assigment
and then used to re-shufled assigment prorpotional to that propensity.

placebo dataset with 
```{r}
dt$cpr <- dt %>%
  add_prop %>%
  select(cpr) %>%
  unlist
summary(dt$start_year)
dt %>%  shuffle_start(2) %>% with(summary(start_year))

#strictly placebo data
set.seed(0203)
placebo_dt <- mclapply(1:200, \(x) {
  start <- Sys.time()
  #compute propensity
  result <- dt %>%  shuffle_start(0) %>%  reset_y0
  result$cpr <- NULL
  end <- Sys.time()
  print(end - start)
  result
}, mc.cores = 20) %>%
  bind_rows(.id = "replication")


#placebo_dt %>% with(tapply(i, list(start_year, replication), first))
#placebo_dt %>% with(tapply(n, list(start_year, replication), mean)) %>% rowMeans
#placebo_dt %>% with(tapply(y0/n*10^5, list(start_year, replication), mean, na.rm = TRUE)) %>% rowMeans

saveRDS(placebo_dt, "data/placebo_dt_prop.rds")
```
