==> analyses/4.1. fit models _new.rmd <==


```{r}
options(digitis = 3)
options(scipen = 10^5) 
options(help_type = "html")
library(data.table)
library(tidyverse)
library(parallel)
#load R functions
sapply(list.files("R/", ".r", full.names = TRUE), source)
set.seed(0203)
options(scipen = 10^6)


library(INLA)
library(fixest)
library(parallel)

source('R/bb etc.r')
```

DATA
```{r}
dt_list <- readRDS(file = paste0("data/placebo_1124.rds"))
length(dt_list)

pre_dat <- \(dat) {
  dat %>%
  mutate(
    Y = round(SMR * E, 0)
    , Y0 = Y
    , Y0 = if_else(O != 1, NA, Y0)
    , id = as.numeric(factor(fid))
    , time = as.numeric(as.factor(fyr))
    , trt = as.numeric(O != 1)
  ) %>%
  group_by(id) %>%
  mutate(
    mxi = mean(x)
    , cohort = min(ifelse(O != 1, time, NA), na.rm = TRUE)
  ) %>%
  group_by(time) %>%
  mutate(mxt = mean(x)) %>%
  ungroup %>%
  mutate(
    etime = time - cohort
    , cohort = if_else(cohort == Inf, 99, cohort)
    , etime = if_else(etime == - Inf, -99, etime)
    , fcohort = factor(cohort)
  )
}

dat <- pre_dat(dt_list[[5]])

nboot <- 40
```

Extended TWFE 
```{r}
library(fixest)
boot_etwfe <- \(nb = nboot) {
  mclapply(1:nb, \(i) {
    bdt <- left_join(
      dat
      , data.frame(id = unique(dat$id), wts = bb(length(unique(dat$id))))
      , by = "id"
    ) %>% mutate(wts = wts / sum(wts) * nrow(dat))
    suppressMessages(
      fepois(
        Y0 ~ 1 | cohort[x] + time[x]
        , vcov = ~ id, offset = ~ log(n)
        , bdt, weights = ~ wts
      ) %>% predict(bdt, type = "reponse")
    )
  }, mc.cores = 20)
}

fit_etwfe <- \(dat, n_boot = nboot) {
  t0 <- Sys.time()
  model <- suppressMessages(
    fepois(Y0 ~ 1 | cohort[x] + time[x]
      , dat, vcov = ~ id, offset = ~ log(n)
    )
  )
  dat$Y0.mean <-  predict(model, dat, type = "reponse")

  dat[, paste0("Y0.b", 1:nboot)] <- boot_etwfe(n_boot)
  dat$Y0.sd   <-  apply(dat[, paste0("Y0.b", 1:n_boot)], 1, sd)

  dat$method <- "etwfe"
  t1 <- Sys.time()
  print(t1 - t0)
  dat
}
fit_etwfe(dat)
```



Pooled Poisson 
```{r}
boot_pois <- \(nb = nboot) {
  mclapply(1:nb, \(i) {
    bdt <- left_join(
      dat
      , data.frame(id = unique(dat$id), wts = bb(length(unique(dat$id))))
      , by = "id"
    ) %>% mutate(wts = wts / sum(wts) * nrow(dat))

    glm(Y0 ~ x + offset(log(n))
      , family = "poisson"
      , bdt, weights = wts
    ) %>% predict(bdt, type = "reponse")
  }, mc.cores = 20)
}

fit_pois <- \(dat, n_boot = nboot) {
  t0 <- Sys.time()
  dat$xdm <- resid(lm(x ~ cohort + time, dat))
  model <- glm(Y0 ~ x + offset(log(n))
    , family = "poisson", dat
  )
  dat$Y0.mean <-  predict(model, dat, type = "reponse")

  dat[, paste0("Y0.b", 1:nboot)] <- boot_pois(n_boot)
  dat$Y0.sd   <-  apply(dat[, paste0("Y0.b", 1:n_boot)], 1, sd)

  dat$method <- "pois"
  t1 <- Sys.time()
  print(t1 - t0)
  dat
}
fit_pois(dat)
```

mixed-effects
```{r}
library(lme4)
boot_me <- \(nb = nboot) {
  mclapply(1:nb, \(i) {
    bdt <- left_join(
      dat
      , data.frame(id = unique(dat$id), wts = bb(length(unique(dat$id))))
      , by = "id"
    ) %>% mutate(wts = wts / sum(wts) * nrow(dat))

    glmer(Y0 ~ offset(log(n)) + x + (x - 1 | id) + (1 | fyr)
      , family = "poisson"
      , bdt, weights = wts
    ) %>% predict(bdt, type = "reponse")
  }, mc.cores = 20)
}

fit_me <- \(dat, n_boot = nboot) {
  t0 <- Sys.time()
  model <- glmer(Y0 ~ offset(log(n)) + x + (x - 1 | id) + (1 | fyr)
    , family = "poisson"
    , dat
  )
  dat$Y0.mean <-  predict(model, dat, type = "reponse")

  dat[, paste0("Y0.b", 1:nboot)] <- boot_me(n_boot)
  dat$Y0.sd   <-  apply(dat[, paste0("Y0.b", 1:n_boot)], 1, sd)

  dat$method <- "me"
  t1 <- Sys.time()
  print(t1 - t0)
  dat
}

fit_me(dat)
```

mixed-effects (conditional on )
```{r}
fit_me_inla <- \(dat, n_boot = nboot) {
  t0 <- Sys.time()
  model <- inla(Y0 ~ x + offset(log(n))
    + f(id, x, model = "iid")
    + f(fyr, model = "iid")
    , family = "poisson"
    , data = dat
    , control.predictor = list(link = 1)
    , control.compute = list(config = TRUE)
  )
  ps <- inla.posterior.sample(n_boot, model)
  y_pred <- exp(inla.posterior.sample.eval("Predictor", ps))
  # y_pred <- apply(y_pred, 1:2, \(l) rpois(1, l))

  dat[, paste0("Y0.b", 1:nboot)] <- y_pred
  dat$Y0.mean   <-  apply(dat[, paste0("Y0.b", 1:n_boot)], 1, mean)
  dat$Y0.sd   <-  apply(dat[, paste0("Y0.b", 1:n_boot)], 1, sd)

  dat$method <- "me_inla"
  t1 <- Sys.time()
  print(t1 - t0)
  dat
}
fit_me_inla(dat)
```

```{r}
fit_sp <- \(dat, n_boot = nboot) {
  t0 <- Sys.time()
  dat$i <- dat$id
  model <- inla(Y0 ~ x + offset(log(n))
    + f(id, x, model = "iid")
    + f(fyr, model = "iid")
    + f(i, x, model = "besag", graph = graph)
    , family = "poisson"
    , data = dat
    , control.predictor = list(link = 1)
    , control.compute = list(config = TRUE)
  )
  ps <- inla.posterior.sample(n_boot, model)
  y_pred <- exp(inla.posterior.sample.eval("Predictor", ps))
  # y_pred <- apply(y_pred, 1:2, \(l) rpois(1, l))

  dat[, paste0("Y0.b", 1:nboot)] <- y_pred
  dat$Y0.mean   <-  apply(dat[, paste0("Y0.b", 1:n_boot)], 1, mean)
  dat$Y0.sd   <-  apply(dat[, paste0("Y0.b", 1:n_boot)], 1, sd)

  dat$method <- "sp"
  t1 <- Sys.time()
  print(t1 - t0)
  dat
}
fit_sp(dat)
```

```{r}
fit_tp <- \(dat, n_boot = nboot) {
  t0 <- Sys.time()
  dat$i <- dat$id
  model <- inla(Y0 ~ x + offset(log(n))
    + f(id, x, model = "iid")
    + f(fyr, model = "iid")
    + f(i, model = "iid", group = time, control.group = list(model = "ar1"))
    , family = "poisson"
    , data = dat
    , control.predictor = list(link = 1)
    , control.compute = list(config = TRUE)
  )
  ps <- inla.posterior.sample(n_boot, model)
  y_pred <- exp(inla.posterior.sample.eval("Predictor", ps))
  # y_pred <- apply(y_pred, 1:2, \(l) rpois(1, l))

  dat[, paste0("Y0.b", 1:nboot)] <- y_pred
  dat$Y0.mean   <-  apply(dat[, paste0("Y0.b", 1:n_boot)], 1, mean)
  dat$Y0.sd   <-  apply(dat[, paste0("Y0.b", 1:n_boot)], 1, sd)

  dat$method <- "tp"
  t1 <- Sys.time()
  print(t1 - t0)
  dat
}
fit_tp(dat)
```

```{r}
fit_st <- \(dat, n_boot = nboot) {
  t0 <- Sys.time()
  dat$i <- dat$id
  model <- inla(Y0 ~ x + offset(log(n))
    + f(id, x, model = "iid")
    + f(fyr, model = "iid")
    + f(i, x, model = "besag", graph = graph
      , group = time, control.group = list(model = "ar1")
    )
    , family = "poisson"
    , data = dat
    , control.predictor = list(link = 1)
    , control.compute = list(config = TRUE)
  )
  ps <- inla.posterior.sample(n_boot, model)
  y_pred <- exp(inla.posterior.sample.eval("Predictor", ps))
  # y_pred <- apply(y_pred, 1:2, \(l) rpois(1, l))

  dat[, paste0("Y0.b", 1:nboot)] <- y_pred
  dat$Y0.mean   <-  apply(dat[, paste0("Y0.b", 1:n_boot)], 1, mean)
  dat$Y0.sd   <-  apply(dat[, paste0("Y0.b", 1:n_boot)], 1, sd)

  dat$method <- "st"
  t1 <- Sys.time()
  print(t1 - t0)
  dat
}
fit_st(dat)
```


```{r}
load("data/gs")
graph <- gs
set.seed(0203)
nboot <- 400

fit_all <- \(dat) {
  etwfe   <- fit_etwfe(dat)
  pois    <- fit_pois(dat)
  me      <- fit_me(dat)
  me_inla <- fit_me_inla(dat)
  sp      <- fit_sp(dat)
  tp      <- fit_tp(dat)
  st      <- fit_st(dat)
  bind_rows(etwfe, pois, me, me_inla, sp, tp, st, .id = "o")
}

colnames(fit_all(dat))

```


```{r}
start_time <- Sys.time()
lapply(names(dt_list), \(name) {
  print(name)
  dt <- dt_list[name][[1]] %>% pre_dat
  rep <-  tryCatch(
    fit_all(dt)
    ,  error = function(c) NULL
  )
  saveRDS(rep, file = paste0("data/tmp/placebo_rep_", name))
})
end_time <- Sys.time()
end_time - start_time
```

```{r}
list_files <- list.files("data/tmp", "placebo_rep_", full.names = TRUE)
list_dat   <- lapply(list_files, \(fl) readRDS(fl))
length(list_dat)
library(data.table)
fits      <- rbindlist(list_dat, idcol = "replication")
saveRDS(fits, file = paste0("data/placebo_1124.rds"))
```
